refactor: useUploader useImageUploader로 네이밍 변경
