+++ src/hooks/index.ts
@@ -1,3 +1,3 @@
 export * from './useMediaQuery'
 export * from './useImage'
-export * from './useUploader'
+export * from './useImageUploader'
